// src/app/shared/interface/theme.interface.ts
export interface Brands {
  status: boolean;
  brand_ids: number[];
}
// ...
export interface Seller {
  status: boolean;
  title: string;
  description?: string;
  store_ids: number[];
}

export interface Sidebar {
  status: boolean;
  categories_icon_list: CategoriesIconList;
  left_side_banners: TwoBanners;
  sidebar_products: SidebarProducts;
}

export interface CategoriesIconList {
  title: string;
  description?: string;
  category_ids: number[];
  status: boolean;
}

export interface TwoBanners {
  status: boolean;
  banner_1: Link;
  banner_2: Link;
}

export interface SidebarProducts {
  title: string;
  product_ids: number[];
  status: boolean;
}

export interface ProductSection {
  title: string;
  description?: string;
  product_ids: number[];
  status: boolean;
}

export interface CategoriesSection {
  title: string;
  description: string;
  image_url: string;
  category_ids: number[];
  status: boolean;
}

export interface FullWidthBanner {
  redirect_link: RedirectLink;
  image_url: string;
  status: boolean;
}

export interface BlogSection {
  title: string;
  description?: string;
  status: boolean;
  blog_ids: number[];
}

export interface NewsLetter {
  title?: string;
  sub_title?: string;
  image_url?: string;
  status?: boolean;
}
// ...
export interface HomeBannerMadrid {
  status: boolean;
  main_banner: Link;
}
// ...
export interface Denver {
  id: number;
  content: DenverContent;
  slug: string;
}

export interface DenverContent {
  home_banner: HomeBannerMadrid;
  categories_icon_list: CategoriesSection;
  products_list_1: ProductSection;
  two_column_banners: TwoBanners;
  slider_product_with_banner: SliderProductWithBanner;
  coupon_banner: FullWidthBanner;
  products_list_2: ProductSection;
  products_list_3: ProductSection;
  news_letter: NewsLetter;
  brands: Brands;
  seller: Seller;
  products_ids: number[];
}

export interface SliderProductWithBanner {
  slider_products: SliderProductsTokyo;
  left_side_banners: RightSideBanners;
}